+++ frontend/src/pages/RoundResultPage/NextRoundButton/NextRoundButton.hook.ts
