+++ frontend/src/pages/RoundResultPage/NextRoundButton/NextRoundButton.tsx
@@ -3,9 +3,9 @@ import { useParams } from 'react-router-dom';
 
 import useMoveNextRoundMutation from './NextRoundButton.hook';
 import { createRandomNextRoundMessage, getNextRoundButtonText } from './NextRoundButton.utils';
-import AlertModal from '../AlertModal/AlertModal';
-import Button from '../Button/Button';
-import { bottomButtonLayout } from '../Button/Button.styled';
+import AlertModal from '../../../components/common/AlertModal/AlertModal';
+import Button from '../../../components/common/Button/Button';
+import { bottomButtonLayout } from '../../../components/common/Button/Button.styled';
 
 import useBalanceContentQuery from '@/hooks/useBalanceContentQuery';
 import useGetUserInfo from '@/hooks/useGetUserInfo';
