+++ frontend/src/pages/RoundResultPage/NextRoundButton/NextRoundButton.utils.ts
