+++ frontend/src/pages/RoundResultPage/RoundVoteContainer/RoundResultTab/RoundResultTab.styled.ts
