+++ frontend/src/pages/RoundResultPage/RoundVoteContainer/RoundVoteContainer.styled.ts
