+++ frontend/src/pages/RoundResultPage/RoundVoteContainer/TabContentContainer/EmptyVoteContent/EmptyVoteContent.tsx
