+++ frontend/src/pages/RoundResultPage/RoundVoteContainer/TabContentContainer/OptionParticipantsContainer/OptionParticipants/NicknameItem/NicknameItem.styled.ts
