+++ frontend/src/pages/RoundResultPage/RoundVoteContainer/TabContentContainer/OptionParticipantsContainer/OptionParticipantsContainer.styled.ts
