+++ frontend/src/pages/RoundResultPage/RoundVoteContainer/TabContentContainer/StatisticBar/StatisticBar.styled.ts
