+++ frontend/src/pages/RoundResultPage/RoundVoteContainer/TabContentContainer/StatisticBar/hooks/useTotalCountAnimation.ts
