+++ frontend/src/pages/RoundResultPage/RoundVoteContainer/TabContentContainer/TabContentContainer.stories.tsx
