+++ frontend/src/pages/RoundResultPage/RoundVoteContainer/TabContentContainer/TabContentContainer.styled.ts
