+++ frontend/src/pages/RoundResultPage/RoundVoteContainer/TabContentContainer/TabContentContainer.util.ts
