+++ frontend/src/pages/RoundResultPage/RoundVoteContainer/TabContentContainer/VoteStatisticContent/VoteStatisticContent.styled.ts
