+++ frontend/src/pages/RoundResultPage/RoundVoteContainer/TabContentContainer/VoteStatisticContent/VoteStatisticContent.tsx
